LPS-140076 Move the color handling logic to the ColorPalette component
