+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/common/components/ColorPalette.js
@@ -15,7 +15,7 @@
 import ClayButton from '@clayui/button';
 import classNames from 'classnames';
 import PropTypes from 'prop-types';
-import React from 'react';
+import React, {useMemo} from 'react';
 
 import {config} from '../../app/config/index';
 import {useId} from '../../app/utils/useId';
@@ -28,32 +28,43 @@ export default function ColorPalette({
 }) {
 	const colorPaletteId = useId();
 
+	const themeColors = useMemo(() => {
+		return config.themeColorsCssClasses.map((color) => {
+			return {
+				color,
+				cssClass: color,
+				rgbValue: getRgbValue(color),
+			};
+		});
+	}, []);
+
 	return (
 		<div className="page-editor__color-palette">
 			{label && <label htmlFor={colorPaletteId}>{label}</label>}
 
 			<div className="palette-container" id={colorPaletteId}>
 				<ul className="list-unstyled palette-items-container">
-					{config.themeColorsCssClasses.map((color) => (
+					{themeColors.map((color) => (
 						<li
 							className={classNames('palette-item', {
 								'palette-item-selected':
-									color === selectedColor,
+									color.rgbValue === selectedColor ||
+									color.cssClass === selectedColor,
 							})}
 							key={color}
 						>
 							<ClayButton
 								block
 								className={classNames(
-									`bg-${color}`,
+									`bg-${color.cssClass}`,
 									'palette-item-inner',
 									'p-1',
 									'rounded-circle'
 								)}
 								displayType="unstyled"
-								onClick={(event) => onColorSelect(color, event)}
+								onClick={() => onColorSelect(color)}
 								small
-								title={color}
+								title={color.cssClass}
 							/>
 						</li>
 					))}
@@ -80,3 +91,18 @@ ColorPalette.propTypes = {
 	onColorSelect: PropTypes.func.isRequired,
 	selectedColor: PropTypes.string,
 };
+
+function getRgbValue(className) {
+	const node = document.createElement('div');
+
+	node.classList.add(`bg-${className}`);
+	node.style.display = 'none';
+
+	document.body.append(node);
+
+	const rgbValue = getComputedStyle(node).backgroundColor;
+
+	document.body.removeChild(node);
+
+	return rgbValue;
+}
